+++ docs/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [2.0.1] - In progress
 ### Added
 - Sphinx style documentation
-- Add pytest and pytest-cov
+- Add tests using pytest and pytest-cov
 - Used [codecov.io](http://www.codecov.io/) to showcase the testcase coverage
 - Bug fixes
 - Code refactoring and improvements